uart staic and buffer review

+++ Proj_atmega128_1/Atmega128/Livraria/Src/buffer.c
@@ -20,7 +20,7 @@ BUFFvar* BUFF_raw(buffer_parameter* par);
 void BUFF_flush(buffer_parameter* par);
 
 /*** Procedure & Function ***/
-BUFF buff_enable( uint8_t size_buff, BUFFvar* buff )
+BUFF buff_enable( uint16_t size_buff, BUFFvar* buff )
 {
 	// OBJECT STRUCT
 	BUFF setup_buffer;
